+++ apps/desktop/src/lib/stack/header/HeaderControlSection.svelte
