+++ apps/desktop/src/lib/stack/header/SeriesRowLabels.svelte
