--- apps/desktop/src/lib/stack/header/index.ts
@@ -1,1 +1,0 @@
-export { default as StackHeader } from './StackHeader.svelte';
